# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Aplikasi Booking Kost API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/transaction-booking-cancel/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction-controller"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Booking",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.BookingSaveResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/transaction-booking/users/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction-controller"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Booking",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.BookingResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/transaction-booking/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction-controller"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Booking",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.BookingResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "transaction-controller"
                ],
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID Kost",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Data Kost",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.BookingSaveRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/common.APIResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/response.BookingSaveResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "common.APIResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "request.BookingSaveRequest": {
            "type": "object",
            "properties": {
                "detailPenghuni": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/request.PenghuniRequest"
                    }
                },
                "jumlahPenghuni": {
                    "type": "integer"
                },
                "kamarId": {
                    "type": "integer"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
...
        "request.PenghuniRequest": {
            "type": "object",
            "properties": {
                "jenisKelamin": {
                    "type": "string"
                },
                "namaPenghuni": {
                    "type": "string"
                },
                "nomorHp": {
                    "type": "integer"
                },
                "nomorKtp": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
        "response.BookingResponse": {
            "type": "object",
            "properties": {
                "detailPenghuni": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.PenghuniResponse"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "jumlahPenghuni": {
                    "type": "integer"
                },
                "namaKost": {
                    "type": "string"
                },
                "statusBooking": {
                    "type": "string"
                },
                "typeKost": {
                    "type": "string"
                }
            }
        },
        "response.BookingSaveResponse": {
            "type": "object",
            "properties": {
                "detailPenghuni": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/response.PenghuniResponse"
                    }
                },
                "jumlahPenghuni": {
                    "type": "integer"
                },
                "namaKost": {
                    "type": "string"
                },
                "statusBooking": {
                    "type": "string"
                }
            }
        },
...
        "response.PenghuniResponse": {
            "type": "object",
            "properties": {
                "jenisKelamin": {
                    "type": "string"
                },
                "namaPenghuni": {
                    "type": "string"
                },
                "nomorHp": {
                    "type": "integer"
                },
                "nomorKtp": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Type \"Bearer \" followed by your JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}